erts: Restructure include of hash.h to work with inline
erl_child_setup is linked together with hash.o so we need to
inline any functions into it when running in debug mode.
